more weapons bam
fixed lot of stuff don't know all of em

+++ Assets/XPDrop.cs
@@ -7,7 +7,7 @@ public class XPDrop : MonoBehaviour
 
     public int xpAmount = 10; // The amount of XP to drop
     public GameObject xpPrefab;
-    public int mobHP;
+    public float mobHP;
     
     void Update()
     {
